backend/requirements.txt
Django==3.2.3
python-dotenv==1.0.0
djangorestframework==3.12.4
PyJWT==2.3.0
gunicorn==20.1.0
psycopg-binary==3.0.10
django-colorfield==0.6.3
psycopg2-binary==2.9.3
djoser==2.1.0
webcolors==1.11.1
drf-extra-fields==3.7.0
django-filter==23.2
reportlab==3.6.9
django-csvimport
bootstrap-colorpicker